Fixed timeout for ScrapingCorvet class

+++ squalaetp/management/commands/importcorvet.py
@@ -70,42 +70,43 @@ class Command(BaseCommand):
 
     def _import_corvet(self, queryset, squalaetp=True, limit=False):
         nb_import = 0
-        scrap = ScrapingCorvet()
-        for query in queryset:
-            start_msg = f"{query.vin}"
-            if squalaetp:
-                start_msg = f"{query.numero_de_dossier} - {start_msg}"
-            start_time = time.time()
-            for attempt in range(2):
-                row = xml_parser(scrap.result(query.vin))
-                if scrap.ERROR or "ERREUR COMMUNICATION SYSTEME CORVET" in row:
-                    nb_import += 1
-                    delay_time = time.time() - start_time
-                    self.stdout.write(
-                        self.style.ERROR(f"{start_msg} error CORVET in {delay_time}"))
+        if queryset:
+            scrap = ScrapingCorvet()
+            for query in queryset:
+                start_msg = f"{query.vin}"
+                if squalaetp:
+                    start_msg = f"{query.numero_de_dossier} - {start_msg}"
+                start_time = time.time()
+                for attempt in range(2):
+                    row = xml_parser(scrap.result(query.vin))
+                    if scrap.ERROR or "ERREUR COMMUNICATION SYSTEME CORVET" in row:
+                        nb_import += 1
+                        delay_time = time.time() - start_time
+                        self.stdout.write(
+                            self.style.ERROR(f"{start_msg} error CORVET in {delay_time}"))
+                        break
+                    elif row and row.get('donnee_date_entree_montage'):
+                        defaults = defaults_dict(Corvet, row, "vin")
+                        obj, created = Corvet.objects.update_or_create(vin=row["vin"], defaults=defaults)
+                        obj.opts.update = False
+                        obj.opts.save()
+                        nb_import += 1
+                        delay_time = time.time() - start_time
+                        self.stdout.write(
+                            self.style.SUCCESS(f"{start_msg} updated in {delay_time}"))
+                        break
+                    if attempt:
+                        if squalaetp:
+                            query.vin_error = True
+                        Corvet.objects.filter(vin=query.vin).delete()
+                        delay_time = time.time() - start_time
+                        self.stdout.write(
+                            self.style.ERROR(f"{start_msg} error VIN in {delay_time}"))
+                if squalaetp:
+                    query.save()
+                if limit and nb_import >= 200:
                     break
-                elif row and row.get('donnee_date_entree_montage'):
-                    defaults = defaults_dict(Corvet, row, "vin")
-                    obj, created = Corvet.objects.update_or_create(vin=row["vin"], defaults=defaults)
-                    obj.opts.update = False
-                    obj.opts.save()
-                    nb_import += 1
-                    delay_time = time.time() - start_time
-                    self.stdout.write(
-                        self.style.SUCCESS(f"{start_msg} updated in {delay_time}"))
-                    break
-                if attempt:
-                    if squalaetp:
-                        query.vin_error = True
-                    Corvet.objects.filter(vin=query.vin).delete()
-                    delay_time = time.time() - start_time
-                    self.stdout.write(
-                        self.style.ERROR(f"{start_msg} error VIN in {delay_time}"))
-            if squalaetp:
-                query.save()
-            if limit and nb_import >= 200:
-                break
-        scrap.close()
+            scrap.close()
         return nb_import
 
     def _import_sivin(self, immat):
